% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_tables.R
\name{ez_differential_group_table}
\alias{ez_differential_group_table}
\title{Easily Table Differential Results from Group Moderated Models}
\usage{
ez_differential_group_table(rep_model, what = "main")
}
\arguments{
\item{rep_model}{The results from one of the ReputationAnalyses
group Models (e.g., rep_auto_group_mod). The model can have between-group equality
constraints; those are marked by missing values (NA) in the group_label column.}

\item{what}{The parameters you want in the table. Current options are
main and all. If what = "main", then just the 'main' model parameters are provided.
This will include, when avaiable, hearsay consensus, hearsay accuracy,
direct accuracy (P1-P1 agreement), P1 Meta-Accuracy, and P2 Meta-Accuracy.}
}
\value{
The function returns an object of class \code{\link[tibble:tbl_df-class]{tbl_df}}.
}
\description{
This takes output from one of the group moderated reputation models
(e.g., rep_analyses_auto) and returns a tibble of differential
(correlational) parameters. It works with any of the group moderated models.
}
\examples{
data("rep_sim_data")
     # Consensus only Model
          agree_rep_consensus_grpmod <- rep_auto_group_mod(data = rep_sim_data,
                                                           p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
                                                           p2_reports = c("B_C_agreeableness", "D_A_agreeableness"),
                                                           group_mod = "study",
                                                           groups_eql = "all",
                                                           params_eql = "all")

          ez_differential_group_table(agree_rep_consensus_grpmod, what = "main")

          ez_differential_group_table(agree_rep_consensus_grpmod, what = "all")

       # Consensus & Accuracy

          agree_rep_con_acc_grpmod <- rep_auto_group_mod(data = rep_sim_data,
                                                         p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
                                                         p2_reports = c("B_C_agreeableness", "D_A_agreeableness"),
                                                         target_self = c("C_C_agreeableness", "A_A_agreeableness"),
                                                         group_mod = "study", groups_eql = "all",params_eql = "all")

          ez_differential_group_table(agree_rep_con_acc_grpmod, what = "main")

          ez_differential_group_table(agree_rep_con_acc_grpmod, what = "all")

      # Consensus, Accuracy, 3rd  Person Meta

agree_full_3pmeta_grpmod <- rep_auto_group_mod(data = rep_sim_data,
                                                      p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
                                                      p2_reports = c("B_C_agreeableness", "D_A_agreeableness"),
                                                      target_self = c("C_C_agreeableness", "A_A_agreeableness"),
                                                      p1_meta = c("A_B_C_agree_meta", "C_D_A_agree_meta"),
                                                      p2_meta = c("B_A_C_agree_meta", "D_C_A_agree_meta"),
                                                      group_mod = "study",
                                                      groups_eql = "all",
                                                      params_eql = "all")

         ez_differential_group_table(agree_full_3pmeta_grpmod, what = "main")

          ez_differential_group_table(agree_full_3pmeta_grpmod, what = "all")
}
