% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_tables.R
\name{ez_elevation_group_table}
\alias{ez_elevation_group_table}
\title{Easily Table Elevation Results from Group Moderated Models}
\usage{
ez_elevation_group_table(rep_model)
}
\arguments{
\item{rep_model}{The results from one of the ReputationAnalyses group moderated models
Models (e.g., rep_auto_group_mod).}
}
\value{
The function returns a list of 2 objects of class \code{\link[tibble:tbl_df-class]{tbl_df}}.
}
\description{
This takes output from one of the reputation group moderated models
(e.g., rep_auto_group_mod) and returns a list of two tibbles corresponding to
elevation (mean-differences) estimates and the relevant pooled Means and SDs respectively.
The elevation table contains the raw difference, 95% CI around the raw difference,
z scores, p values, and a cohen's d (difference / pooled SD). It has only 1 parameter.
}
\examples{
data("rep_sim_data")

       # Consensus only Model
          agree_rep_consensus_grpmod <- rep_auto_group_mod(data = rep_sim_data,
                                                           p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
                                                           p2_reports = c("B_C_agreeableness", "D_A_agreeableness"),
                                                           group_mod = "study",
                                                           groups_eql = "all",
                                                           params_eql = "all")

          ez_elevation_group_table(agree_rep_consensus_grpmod)

          ez_elevation_group_table(agree_rep_consensus_grpmod)

       # Consensus & Accuracy

          agree_rep_con_acc_grpmod <- rep_auto_group_mod(data = rep_sim_data,
                                                         p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
                                                         p2_reports = c("B_C_agreeableness", "D_A_agreeableness"),
                                                         target_self = c("C_C_agreeableness", "A_A_agreeableness"),
                                                         group_mod = "study", groups_eql = "all",params_eql = "all")

          ez_elevation_group_table(agree_rep_con_acc_grpmod)

          ez_elevation_group_table(agree_rep_con_acc_grpmod)

      # Consensus, Accuracy, 3rd  Person Meta

agree_full_3pmeta_grpmod <- rep_auto_group_mod(data = rep_sim_data,
                                                      p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
                                                      p2_reports = c("B_C_agreeableness", "D_A_agreeableness"),
                                                      target_self = c("C_C_agreeableness", "A_A_agreeableness"),
                                                      p1_meta = c("A_B_C_agree_meta", "C_D_A_agree_meta"),
                                                      p2_meta = c("B_A_C_agree_meta", "D_C_A_agree_meta"),
                                                      group_mod = "study",
                                                      groups_eql = "all",
                                                      params_eql = "all")

         ez_elevation_group_table(agree_full_3pmeta_grpmod)

          ez_elevation_group_table(agree_full_3pmeta_grpmod)

}
