% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reputation_model.R
\name{ez_elevation_table}
\alias{ez_elevation_table}
\title{Easily Table Elevation Results}
\usage{
ez_elevation_table(rep_model)
}
\arguments{
\item{rep_model}{The results from one of the ReputationAnalyses
Models (e.g., rep_analyses_auto).}
}
\value{
The function returns a list of 2 objects of class \code{\link[tibble::tibble()]{tibble}}.
}
\description{
This takes output from one of the reputational analysis models
(e.g., rep_analyses_auto) and returns a list of two tibbles corresponding to
elevation (mean-differences) estimates and the relevant pooled Means and SDs respectively.
The elevation table contains the raw difference, 95% CI around the raw difference,
z scores, p values, and a cohen's d (difference / pooled SD). It has only 1 parameter.
}
\examples{
data("rep_sim_data")

       # Consensus only Model
          agree_rep_consensus <- rep_analyses_auto(data = rep_sim_data,
                        p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
                        p2_reports = c("B_C_agreeableness", "D_A_agreeableness"))

          ez_elevation_table(agree_rep_consensus)

       # Consensus & Accuracy

          agree_rep_con_acc <- rep_analyses_auto(data = rep_sim_data,
                       p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
                       p2_reports = c("B_C_agreeableness", "D_A_agreeableness"),
                       target_self = c("C_C_agreeableness", "A_A_agreeableness"))

          ez_elevation_table(agree_rep_con_acc)

      # Consensus, Accuracy, 3rd  Person Meta

         agree_rep_all <- rep_analyses_auto(data = rep_sim_data,
                       p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
                       p2_reports = c("B_C_agreeableness", "D_A_agreeableness"),
                       target_self = c("C_C_agreeableness", "A_A_agreeableness"),
                       p1_meta = c("A_B_C_agree_meta", "C_D_A_agree_meta"),
                       p2_meta = c("B_A_C_agree_meta", "D_C_A_agree_meta"))

         ez_elevation_table(agree_rep_all)

}
