% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_tables.R
\name{ez_group_id_mod_table}
\alias{ez_group_id_mod_table}
\title{Easily Table Model with Group and Individual Difference Moderators}
\usage{
ez_group_id_mod_table(rep_model)
}
\arguments{
\item{rep_model}{The results from one of the ReputationAnalyses
group- and individual-level moderator Models (e.g., rep_accuracy_group_id_mods).}
}
\value{
The function returns an object of class \code{\link[tibble::tibble()]{tibble}}.
}
\description{
This takes output from one of the group- and individual-level moderated reputation models
(e.g., rep_accuracy_group_id_mods) and returns a tibble of model
(regression) parameters. It works with any of the Group- and Individual-level moderated models.
}
\examples{
data("rep_sim_data")
moderator_data <- rep_sim_data \%>\%
dplyr::mutate(B_C_agreeableness_cent = scale(B_C_agreeableness, scale = FALSE),
       D_A_agreeableness_cent = scale(D_A_agreeableness, scale = FALSE),
       B_iri_perspective_cent = scale(B_iri_perspective, scale = FALSE),
       D_iri_perspective_cent = scale(D_iri_perspective, scale = FALSE),
       B_ptXagree_interaction = B_C_agreeableness_cent*B_iri_perspective_cent,
       D_ptXagree_interaction = D_A_agreeableness_cent*D_iri_perspective_cent)

# Example for hearsay accuracy with no equality constraints
agree_pt_mod_fit <- rep_accuracy_group_id_mods(moderator_data,
                                          target_self = c("C_C_agreeableness", "A_A_agreeableness"),
                                          p2_reports = c("B_C_agreeableness_cent", "D_A_agreeableness_cent"),
                                          id_mod_variable = c("B_iri_perspective_cent", "D_iri_perspective_cent"),
                                          interaction_term = c("B_ptXagree_interaction", "D_ptXagree_interaction"),
                                          group_mod = "study")
 ez_group_id_mod_table(agree_pt_mod_fit)

 # Example for hearsay accuracy with all parameters equal across all groups
 agree_pt_mod_fit_alleq <- rep_accuracy_group_id_mods(moderator_data,
                                          target_self = c("C_C_agreeableness", "A_A_agreeableness"),
                                          p2_reports = c("B_C_agreeableness_cent", "D_A_agreeableness_cent"),
                                          id_mod_variable = c("B_iri_perspective_cent", "D_iri_perspective_cent"),
                                          interaction_term = c("B_ptXagree_interaction", "D_ptXagree_interaction"),
                                          group_mod = "study", groups_eql = "all", params_eql = "all")
 ez_group_id_mod_table(agree_pt_mod_fit_alleq)

 # Example with some equality constraints

 agree_pt_mod_fit_someeql <-  rep_accuracy_group_id_mods(moderator_data,
                                                         target_self = c("C_C_agreeableness", "A_A_agreeableness"),
                                                         p2_reports = c("B_C_agreeableness_cent", "D_A_agreeableness_cent"),
                                                         id_mod_variable = c("B_iri_perspective_cent", "D_iri_perspective_cent"),
                                                         interaction_term = c("B_ptXagree_interaction", "D_ptXagree_interaction"),
                                                         group_mod = "group_var", groups_eql = c(1, 4), params_eql = "all")

ez_group_id_mod_table(agree_pt_mod_fit_someeql)
}
