% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_mods.R
\name{rep_con_acc_group_mod_builder}
\alias{rep_con_acc_group_mod_builder}
\title{Reputation Consensus & Accuracy with Group Moderator Model Builder}
\usage{
rep_con_acc_group_mod_builder(p1_reports, p2_reports, target_self,
  groups = NULL, n_triads = length(p1_reports), n_p1s_per_p2s = 1,
  n_p2s_per_p1s = 1)
}
\arguments{
\item{p1_reports}{Quoted column names that contain P1 reports,
or ratings made by the person that knows the target directly.
If more than one is supplied, the target-wise order must match the other
rating types.}

\item{p2_reports}{Quoted column names that contain P2 reports,
or ratings made by the person that knows the target indirectly through the corresponding P1.
If more than one is supplied, the target-wise order must match the other
rating types.}

\item{target_self}{Quoted column names that contain target self-reports.
If more than one is supplied, the target-wise order must match the other
rating types.}

\item{groups}{Vector of quoted group labels (from group-level categorical moderator).}

\item{n_triads}{The number of exchangeable triads in each group. By default, this is determined by
counting the number of P1 reports. It is rare that this parameter would need to be changed.}

\item{n_p1s_per_p2s}{The number of P1s for every P2. This defaults to 1.
Currently, only values of 1 are supported.}

\item{n_p2s_per_p1s}{The number of P2s for every P1;. This defaults to 1.
Currently, only values of 1 are supported.}

\item{data}{The dataframe.
Data should be wide, with a row for every group of participants.
At a minimum, it must contain three columns: one for P1-reports, one for P2-reports, and
one for targets' self-ratings.}

\item{model}{Optional. A model from the corresponding ReputationAnalyses model builder function. If this
is supplied, no additional arguments need to be specified.}

\item{n_p1s_per_ts}{The number of P1s for every target;. This defaults to 1.
Currently, only values of 1 are supported.}

\item{n_p2s_per_ts}{The number of P2s for every target;. This defaults to 1.
Currently, only values of 1 are supported.}

\item{n_ts_per_p1s}{The number of targets for every P1;. This defaults to 1.
Currently, only values of 1 are supported.}

\item{n_ts_per_p2s}{The number of targets for every P2;. This defaults to 1.
Currently, only values of 1 are supported.}
}
\value{
The function returns a list containing an
object of class \code{\link[tibble:tibble-class]{tibble}} with model information
and a string object of the model in lavaan syntax. Model information
includes the type of model, the number of exchangeable triads, and the number
of p1s per p2s, and the number of p2s per p1s.
}
\description{
This builds a model of lavaan syntax for estimating the possible hearsay reputation parameters
as a multi-group path model given vectors of P1, P2, and target self-reports
(vectors of quoted variable names) and a group-level categorical variable.
The baseline model estimates each parameter seperately,
labelling parameters based on the labels of moderator.
Those parameters are:
\describe{
\item{hc}{hearsay consensus; the correlation between P1(T) & P2(T)}
\item{ha}{hearsay accuracy; the correation between P2(T) & T(T)}
\item{da}{direct accuracy; the correlation between P1(T) & T(T)}
\item{int_p1}{Intercept for P1(T)}
\item{int_p2}{Intercept for P2(T)}
\item{int_self}{Intercept for T(T)}
\item{v_p1}{variance for P1(T)}
\item{v_p2}{variance for P2(T)}
\item{v_self}{variance for T(T)}
\item{p1_p2_rel_el}{P1-P2 Relative Elevation (i.e., Mean P1(T) - Mean P2(T))}
\item{self_p2_rel_el}{Self-P2 Relative Elevation (i.e., Mean T(T) - Mean P2(T))}
\item{self_p1_rel_el}{Self-P1 Relative Elevation (i.e., Mean T(T) - Mean P1(T))}
}
\emph{If n exchangeable triads > 1:}
\describe{
\item{rec}{direct reciprocity; the correlation between opposit P1(T)s (e.g., A(C) <-> C(A))}
\item{h}{hearsay reciprocity; the correlation between exchangeable P2(T)s (e.g., B(C) <-> D(A))}
\item{m}{unnamed parameter; The correlation between P2(T) and the opposite P1(T) in a group. (e.g., B(C) <-> C(A))}
\item{tru_sim}{True Similarity; the correlation between targets' self-reports. (e.g., A(A) <-> C(C))}
\item{as_sim_3p}{Third-person assumed similarity; correlation between P2(T) and P1's self-report (e.g., B(C) <- A(A))}
\item{as_sim_1p}{First-person assumed similarity (i.e., interpersonal assumed similarity); correlation betweenP1(T) and P1's self-report (e.g., A(C) <-> A(A))}
}
The function can handle up to n exchangeable triads.
}
\examples{
 # build the model
  agree_con_acc_model_grpmod <- rep_con_acc_group_mod_builder(p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
  p2_reports = c("B_C_agreeableness", "D_A_agreeableness"),
  target_self = c("C_C_agreeableness", "A_A_agreeableness"),
  groups = c("Study_1", "Study_2"))

 # view the model
 agree_consensus_model_grpmod$model

 # view the model information
 # agree_consensus_model_grpmod$rep_model_info

}
