% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_mods.R
\name{rep_consensus_group_mod_builder}
\alias{rep_consensus_group_mod_builder}
\title{Reputation Consensus with Group Moderator Model Builder}
\usage{
rep_consensus_group_mod_builder(p1_reports, p2_reports, groups = NULL,
  use_labs = TRUE, n_triads = length(p1_reports), n_p1s_per_p2s = 1,
  n_p2s_per_p1s = 1)
}
\arguments{
\item{p1_reports}{Quoted column names that contain P1 reports,
or ratings made by the person that knows the target directly.
If more than one is supplied, the target-wise order must match the other
rating types.}

\item{p2_reports}{Quoted column names that contain P2 reports,
or ratings made by the person that knows the target indirectly through the corresponding P1.
If more than one is supplied, the target-wise order must match the other
rating types.}

\item{groups}{Vector of quoted group labels (from group-level categorical moderator).}

\item{use_labs}{Logical indicating whether or not to use the group labels to create the parameter labels.
If FALSE, generic labels (grp1 to grpk, where k is the number of groups) are used.}

\item{n_triads}{The number of exchangeable triads in each group. By default, this is determined by
counting the number of P1 reports. This parameter rarely needs to be changed.}

\item{n_p1s_per_p2s}{The number of P1s for every P2. This defaults to 1.
Currently, only values of 1 are supported.}

\item{n_p2s_per_p1s}{The number of P2s for every P1;. This defaults to 1.
Currently, only values of 1 are supported.}
}
\value{
The function returns a list containing an
object of class \code{\link[tibble:tibble-class]{tibble}} with model information
and a string object of the model in lavaan syntax. Model information
includes the type of model, the number of exchangeable triads, and the number
of p1s per p2s, and the number of p2s per p1s.
}
\description{
This builds a model of lavaan syntax for estimating the possible hearsay reputation parameters
as a multi-group path model given vectors of P1 and P2 reports (vectors of quoted variable names)
and a group-level categorical variable.
The baseline model estimates each parameter seperately,
labelling parameters based on the labels of moderator.
Those parameters are:
\describe{
\item{hc}{hearsay consensus; the correlation between P1(T) & P2(T)}
\item{int_p1}{Intercept for P1(T)}
\item{int_p2}{Intercept for P2(T)}
\item{v_p1}{variance for P1(T)}
\item{v_p2}{variance for P2(T)}
\item{p1_p2_rel_el}{P1-P2 Relative Elevation (i.e., Mean P1(T) - Mean P2(T))}
}
\emph{If n exchangeable triads > 1:}
\describe{
\item{rec}{direct reciprocity; the correlation between opposit P1(T)s (e.g., A(C) <-> C(A))}
\item{h}{hearsay reciprocity; the correlation between exchangeable P2(T)s (e.g., B(C) <-> D(A))}
\item{m}{unnamed parameter; The correlation between P2(T) and the opposite P1(T) in a group. (e.g., B(C) <-> C(A))}
}
The function can handle up to n exchangeable triads.
}
\examples{
 # build the model
  agree_consensus_model_grpmod <- rep_consensus_group_mod_builder(p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
                                                                  p2_reports = c("B_C_agreeableness", "D_A_agreeableness"),
                                                                  groups = c("Study_1", "Study_2"))

 # view the model
 agree_consensus_model_grpmod$model

 # view the model information
  agree_consensus_model_grpmod$rep_model_info

}
