% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reputation_model.R
\name{rep_id_mods_accuracy}
\alias{rep_id_mods_accuracy}
\title{Individual level Moderators of Hearsay Accuracy}
\usage{
rep_id_mods_accuracy(data, target_self, p2_reports, id_mod_variable,
  interaction_term, n_triads = length(target_self), n_ts_per_p2s = 1,
  n_p2s_per_ts = 1)
}
\arguments{
\item{data}{The dataframe that contains the ratings, moderator variable, and the interaction term.
Data should be wide, with a row for every group of participants.
At a minimum, it must contain four columns: one for target self-reports, one for mean-centered P2 reports,
one for the mean-centered moderator variable, and one for the interaction term.}

\item{target_self}{The column(s) that contain target self-reports.
If more than one is supplied, the order must match the other
rating types.}

\item{p2_reports}{The column(s) that contain P2 reports,
or ratings made by the person that knows the target indirectly through the corresponding P1.
Ratings should be grand-mean-centered to increase the interpretibility of the model parameters.
If more than one is supplied, the target-wise order must match the other rating types.}

\item{id_mod_variable}{The column(s) that contain the individual-level moderator of interest.
If more than one is supplied from multiple exchangeable triads,
the order must match the order of the ratings. Like P2-reports, the variable should be mean-centered
to facilitate interpretability.}

\item{interaction_term}{The column(s) that contain the interaction term, or the product of the
mean-centered P2-report and the mean-centered moderator variable. If more than one is supplied from multiple
exchangeable triads, the target-wise order must match the order of the ratings.}

\item{n_triads}{The number of exchangeable triads in each group. By default, this is determined by
counting the number of target self-reports. This parameter rarely needs to be changed.}

\item{n_ts_per_p2s}{The number of targets that each P2 rated. This defaults to 1.
Currently, only values of 1 are supported.}

\item{n_p2s_per_ts}{The number of P2s that rated each target;. This defaults to 1.
Currently, only values of 1 are supported.}
}
\value{
The function returns an object of class \code{\link[lavaan:lavaan-class]{lavaan}}.
}
\description{
This takes the variables needed to assess an individual-level moderator on hearsay acuracy, and fits
a model estimating the corresponding parameters. At a minimum, it requires target self-reports, P2-reports,
an individual-level moderator, and the interaction term.
Note that the P2-reports and moderator variable should be mean-centered.
}
\details{
The parameters for the individual-level moderator analyses are:

\describe{
\item{ha_me}{hearsay accuracy main effect; this should correspond to hearsay accuracy at average
level of moderator variable (if data were properly mean-centered).}
\item{mod_me}{The meain effect of the moderator variable; it can be interpreted as the difference in
target self-reports related to differences in the individual-level moderator variable.}
\item{interaction}{This is the interaction term. It indicates the extent to which hearsay accuracy,
depends on the moderator variable}
\item{v_t}{variance for T(T)}
\item{v_p2}{variance for P2(T)}
\item{v_mod}{variance for moderator variable}
\item{v_interaction}{variance for interaction term}
\item{int_t}{intercept for T(T)}
\item{int_p2}{intercept for P2(T)}
\item{int_mod}{intercept for moderator variable}
\item{int_interaction}{intercept for interaction term}
}
The function can handle up to n exchangeable triads.
}
\examples{
data("rep_sim_data")
          # Prepare data
          library(tidyverse)
         moderator_data <- rep_sim_data \%>\%
           mutate(B_C_agreeableness_cent = scale(B_C_agreeableness, scale = FALSE),
                  D_A_agreeableness_cent = scale(D_A_agreeableness, scale = FALSE),
                  B_iri_perspective_cent = scale(B_iri_perspective, scale = FALSE),
                  D_iri_perspective_cent = scale(D_iri_perspective, scale = FALSE),
                  B_ptXagree_interaction = B_C_agreeableness_cent*B_iri_perspective_cent,
                  D_ptXagree_interaction = D_A_agreeableness_cent*D_iri_perspective_cent)

          rep_id_mods_accuracy(data = moderator_data,
                                target_self = c("C_C_agreeableness", "A_A_agreeableness"),
                                p2_reports = c("B_C_agreeableness_cent", "D_A_agreeableness_cent"),
                                id_mod_variable = c("B_iri_perspective_cent", "D_iri_perspective_cent"),
                                interaction_term = c("B_ptXagree_interaction", "D_ptXagree_interaction"))

}
