% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reputation_model.R
\name{rep_id_mods_generic}
\alias{rep_id_mods_generic}
\title{Individual level Moderators (Generic)}
\usage{
rep_id_mods_generic(data, rating_1, rating_2, id_mod_variable, interaction_term,
  n_triads = length(rating_1), n_r1_per_r2 = 1, n_r2_per_r1 = 1)
}
\arguments{
\item{data}{The dataframe that contains the ratings, moderator variable, and the interaction term.
Data should be wide, with a row for every group of participants.
At a minimum, it must contain four columns: two for ratings (of the same target),
one for the mean-centered moderator variable, and one for the interaction term.}

\item{rating_1}{The column(s) that contain the first rating variable. This might be P1 reports
if investigating moderation of hearsay consensus or self-reports for moderation of hearsay accuracy.
If more than one is supplied, the target-wise order must match across variables.}

\item{rating_2}{The column(s) that second rating variable. For hearsay consensus or accuracy,
this would be P2 reports. If more than one is supplied, the target-wise order must match across variables.}

\item{id_mod_variable}{The column(s) that contain the individual-level moderator of interest.
If more than one is supplied from multiple exchangeable dyads/triads,
the order must match the order of the ratings. Like P2-reports, the variable should be mean-centered
to facilitate interpretability.}

\item{interaction_term}{The column(s) that contain the interaction term, or the product of the
mean-centered P2-report and the mean-centered moderator variable. If more than one is supplied from multiple
exchangeable dyads/triads, the target-wise order must match the order of the ratings.}

\item{n_triads}{The number of exchangeable triads in each group. By default, this is determined by
counting the number of P1 reports. This parameter rarely needs to be changed.}

\item{n_r1_per_r2}{The number of first ratings for each second rating. Currently, only 1:1 is supported.}

\item{n_r1_per_r2}{The number of second ratings for each first rating. Currently, only 1:1 is supported.}
}
\value{
The function returns an object of class \code{\link[lavaan:lavaan-class]{lavaan}}.
}
\description{
This is a generic function for individual-level moderators on two distinguishable ratings on the same target.
This could be P1- and P2- reports, P2- and self-reports, P1- and self-reports, or any other sets of
distinguishable ratings.
}
\details{
The parameters for the individual-level moderator analyses are:

\describe{
\item{rating_me}{main effect of other rating; this should correspond to correlation between ratings at average
level of moderator variable (if data were properly mean-centered).}
\item{mod_me}{The meain effect of the moderator variable; it can be interpreted as the difference in
rating_1 to differences in the individual-level moderator variable.}
\item{interaction}{This is the interaction term. It indicates the extent to which
the correlation between ratings depends on the moderator variable}
\item{v_rating_1}{variance for first rating}
\item{v_rating_2}{variance for second rating}
\item{v_mod}{variance for moderator variable}
\item{v_interaction}{variance for interaction term}
\item{int_rating_1}{intercept for first rating}
\item{int_rating_2}{intercept for second rating}
\item{int_mod}{intercept for moderator variable}
\item{int_interaction}{intercept for interaction term}
}
The function can handle up to n exchangeable triads.
}
\examples{
data("rep_sim_data")
          # Prepare data
          library(tidyverse)
         moderator_data <- rep_sim_data \%>\%
           mutate(B_C_agreeableness_cent = scale(B_C_agreeableness, scale = FALSE),
                  D_A_agreeableness_cent = scale(D_A_agreeableness, scale = FALSE),
                  B_iri_perspective_cent = scale(B_iri_perspective, scale = FALSE),
                  D_iri_perspective_cent = scale(D_iri_perspective, scale = FALSE),
                  B_ptXagree_interaction = B_C_agreeableness_cent*B_iri_perspective_cent,
                  D_ptXagree_interaction = D_A_agreeableness_cent*D_iri_perspective_cent)

          rep_id_mods_generic(data = moderator_data,
                                p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
                                p2_reports = c("B_C_agreeableness_cent", "D_A_agreeableness_cent"),
                                id_mod_variable = c("B_iri_perspective_cent", "D_iri_perspective_cent"),
                                interaction_term = c("B_ptXagree_interaction", "D_ptXagree_interaction"))

}
