% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{rep_model-class}
\alias{rep_model-class}
\alias{rep_model}
\title{An S4 class to represent a reputation models. It
contains (i.e., inherits) everything from the lavaan class.}
\description{
An S4 class to represent a reputation models. It
contains (i.e., inherits) everything from the lavaan class.
}
\section{Slots}{

\describe{
\item{\code{model_type}}{A character string containing the type of model.}
}}

\seealso{
\linkS4class{lavaan}
}
